split types from amidar
