+++ tb_amidar/src/amidar.rs
@@ -5,6 +5,7 @@ use toybox_core;
 use toybox_core::graphics::{Color, Drawable, FixedSpriteData};
 use toybox_core::random;
 use toybox_core::{AleAction, Direction, Input, QueryError};
+use types::*;
 
 use rand::seq::SliceRandom;
 
@@ -90,33 +91,6 @@ mod inits {
     pub const ENEMY_STARTING_SPEED: i32 = 10;
     pub const PLAYER_SPEED: i32 = 8;
 }
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Amidar {
-    pub rand: random::Gen,
-    board: Vec<String>,
-    pub player_start: TilePoint,
-    bg_color: Color,
-    player_color: Color,
-    unpainted_color: Color,
-    painted_color: Color,
-    enemy_color: Color,
-    inner_painted_color: Color,
-    start_lives: i32,
-    start_jumps: i32,
-    render_images: bool,
-    chase_time: i32,
-    chase_score_bonus: i32,
-    jump_time: i32,
-    box_bonus: i32,
-    /// This should be false if you ever use a non-default board.
-    default_board_bugs: bool,
-    enemies: Vec<MovementAI>,
-    level: i32,
-    /// How many previous junctions should the player and enemies remember?
-    history_limit: u32,
-    enemy_starting_speed: i32,
-    player_speed: i32,
-}
 
 impl Amidar {
     pub fn colors(&self) -> Vec<&Color> {
@@ -166,11 +140,6 @@ impl Default for Amidar {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct ScreenPoint {
-    pub sx: i32,
-    pub sy: i32,
-}
 impl ScreenPoint {
     fn new(sx: i32, sy: i32) -> ScreenPoint {
         ScreenPoint { sx, sy }
@@ -180,12 +149,6 @@ impl ScreenPoint {
     }
 }
 
-/// Strongly-typed vector for "world" positioning in Amidar.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct WorldPoint {
-    pub x: i32,
-    pub y: i32,
-}
 impl WorldPoint {
     fn new(x: i32, y: i32) -> WorldPoint {
         WorldPoint { x, y }
@@ -209,12 +172,6 @@ impl WorldPoint {
     }
 }
 
-/// Strongly-typed vector for "tile" positioning in Amidar.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct TilePoint {
-    pub tx: i32,
-    pub ty: i32,
-}
 impl TilePoint {
     pub fn new(tx: i32, ty: i32) -> TilePoint {
         TilePoint { tx, ty }
@@ -232,14 +189,6 @@ impl TilePoint {
         let (dx, dy) = dir.delta();
         self.translate(dx, dy)
     }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct GridBox {
-    pub top_left: TilePoint,
-    pub bottom_right: TilePoint,
-    pub painted: bool,
-    pub triggers_chase: bool,
 }
 
 impl GridBox {
@@ -288,13 +237,6 @@ impl GridBox {
     }
 }
 
-#[derive(Clone, Copy, PartialEq, Eq, Debug, Serialize, Deserialize)]
-pub enum Tile {
-    Empty,
-    Unpainted,
-    ChaseMarker,
-    Painted,
-}
 impl Tile {
     fn new_from_char(c: char) -> Result<Tile, String> {
         match c {
@@ -319,36 +261,6 @@ impl Tile {
     }
 }
 
-#[derive(Clone, PartialEq, Serialize, Deserialize, Debug)]
-pub enum MovementAI {
-    Player,
-    EnemyLookupAI {
-        next: u32,
-        default_route_index: u32,
-    },
-    EnemyPerimeterAI {
-        start: TilePoint,
-    },
-    EnemyAmidarMvmt {
-        vert: Direction,
-        horiz: Direction,
-        start_vert: Direction,
-        start_horiz: Direction,
-        start: TilePoint,
-    },
-    EnemyRandomMvmt {
-        start: TilePoint,
-        start_dir: Direction,
-        dir: Direction,
-    },
-    EnemyTargetPlayer {
-        start: TilePoint,
-        start_dir: Direction,
-        vision_distance: i32,
-        dir: Direction,
-        player_seen: Option<TilePoint>,
-    },
-}
 
 impl MovementAI {
     /// Resetting the mob AI state after player death.
@@ -557,16 +469,6 @@ impl MovementAI {
     }
 }
 
-/// Mob is a videogame slang for "mobile" unit. Players and Enemies are the same struct.
-#[derive(Clone, Serialize, Deserialize)]
-pub struct Mob {
-    pub ai: MovementAI,
-    pub position: WorldPoint,
-    pub caught: bool,
-    speed: i32,
-    step: Option<TilePoint>,
-    history: VecDeque<u32>,
-}
 impl Mob {
     fn new(ai: MovementAI, position: WorldPoint, speed: i32) -> Mob {
         Mob {
@@ -706,24 +608,6 @@ lazy_static! {
         .collect();
 }
 
-#[derive(Clone, Serialize, Deserialize)]
-pub struct Board {
-    pub tiles: Vec<Vec<Tile>>,
-    pub width: u32,
-    pub height: u32,
-    pub junctions: HashSet<u32>,
-    pub chase_junctions: HashSet<u32>,
-    pub boxes: Vec<GridBox>,
-}
-
-#[derive(Clone, PartialEq, Eq, Serialize, Deserialize, Debug)]
-pub struct BoardUpdate {
-    pub vertical: i32,
-    pub horizontal: i32,
-    pub num_boxes: i32,
-    pub triggers_chase: bool,
-    pub junctions: Option<(u32, u32)>,
-}
 impl BoardUpdate {
     fn new() -> BoardUpdate {
         BoardUpdate {
@@ -1096,25 +980,6 @@ impl Board {
     }
 }
 
-#[derive(Clone, Serialize, Deserialize)]
-pub struct StateCore {
-    pub rand: random::Gen,
-    pub score: i32,
-    pub lives: i32,
-    pub jumps: i32,
-    pub chase_timer: i32,
-    pub jump_timer: i32,
-    pub player: Mob,
-    pub enemies: Vec<Mob>,
-    pub board: Board,
-    pub level: i32,
-}
-
-pub struct State {
-    pub config: Amidar,
-    pub state: StateCore,
-}
-
 impl State {
     pub fn try_new(config: &Amidar) -> Result<State, String> {
         let board = Board::try_new(&config.board)?;
@@ -1192,12 +1057,6 @@ impl State {
     }
 }
 
-#[derive(PartialEq, Eq, Clone, Copy)]
-enum EnemyPlayerState {
-    Miss,
-    PlayerDeath,
-    EnemyCatch(usize),
-}
 
 impl toybox_core::Simulation for Amidar {
     fn reset_seed(&mut self, seed: u32) {
